Here's the math needed for the shrink/enlarge undo actions that was left out from the previous commit.
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@1510 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/scribusview.cpp
@@ -5213,7 +5213,7 @@ void ScribusView::MirrorPolyV()
 	MarkClip(b);
 }
 
-void ScribusView::TransformPoly(int mode, int rot, int scaling)
+void ScribusView::TransformPoly(int mode, int rot, double scaling)
 {
 	PageItem *b = SelItem.at(0);
 	b->ClipEdited = true;
